refactor(mixin): gather the Sodium mixins in their own package
Every Sodium bump rechecks each mixin target against the jar the build
resolves, and the list of what to recheck was a grep for net.caffeinemc
across the whole mixin package. In mixin/sodium/ it is a directory
listing. The two accessors on Sodium types stay with the other accessors,
which is where the engine imports them from.
