+++ common/src/main/java/dev/vitrail/mixin/sodium/BlockRendererMixin.java
@@ -1,4 +1,4 @@
-package dev.vitrail.mixin;
+package dev.vitrail.mixin.sodium;
 
 import dev.vitrail.sodium.TerrainVertex;
 import dev.vitrail.render.BlockStateIds;
