+++ common/src/main/java/dev/vitrail/mixin/sodium/MixinDefaultChunkRenderer.java
@@ -1,4 +1,4 @@
-package dev.vitrail.mixin;
+package dev.vitrail.mixin.sodium;
 
 import dev.vitrail.sodium.SodiumPasses;
 import dev.vitrail.pack.program.TerrainPass;
