+++ common/src/main/resources/vitrail.mixins.json
@@ -7,17 +7,17 @@
 		"defaultRequire": 1
 	},
 	"client": [
-		"ArenaAggregatorMixin",
+		"sodium.ArenaAggregatorMixin",
 		"access.BiomeClimateAccessor",
 		"BiomeMixin",
 		"BiomesMixin",
 		"BlockEntityRenderDispatcherMixin",
 		"access.BlockEntityRenderStateAccessor",
-		"BlockRendererMixin",
+		"sodium.BlockRendererMixin",
 		"BufferBuilderMixin",
 		"access.ByteBufferBuilderAccessor",
-		"ChunkMeshFormatsMixin",
-		"ChunkVertexMixin",
+		"sodium.ChunkMeshFormatsMixin",
+		"sodium.ChunkVertexMixin",
 		"ClientLevelMixin",
 		"CloudRendererMixin",
 		"access.CommandEncoderAccessor",
@@ -30,8 +30,8 @@
 		"access.IntermediaryShaderModuleAccessor",
 		"IntermediaryShaderModuleMixin",
 		"MinecraftStartupGuardMixin",
-		"MixinDefaultChunkRenderer",
-		"DefaultFluidRendererMixin",
+		"sodium.MixinDefaultChunkRenderer",
+		"sodium.DefaultFluidRendererMixin",
 		"EntityRenderDispatcherMixin",
 		"FlameFeatureRendererMixin",
 		"GameRendererBlurMixin",
@@ -59,18 +59,18 @@
 		"RenderPassMixin",
 		"access.RenderPipelineAccessor",
 		"RenderPipelineMixin",
-		"MixinRenderRegion",
+		"sodium.MixinRenderRegion",
 		"access.RenderSectionManagerAccessor",
 		"access.RenderTargetAccessor",
 		"RenderTypeFeatureRendererGroupMixin",
 		"RenderTypeFeatureRendererMixin",
-		"MixinShaderChunkRenderer",
+		"sodium.MixinShaderChunkRenderer",
 		"SkyRendererMixin",
 		"access.MixinSodiumWorldRenderer",
-		"MixinSodiumWorldRendererInit",
+		"sodium.MixinSodiumWorldRendererInit",
 		"StagedVertexBufferDrawMixin",
 		"access.TextureManagerAccessor",
-		"VKDrawContextMixin",
+		"sodium.VKDrawContextMixin",
 		"VulkanBackendMixin",
 		"VulkanBindGroupLayoutMixin",
 		"access.VulkanCommandEncoderAccessor",
